\name{autodiff2r-package}
\alias{autodiff2r-package}
\alias{autodiff2r}
\docType{package}
\title{\packageTitle{autodiff2r}}
\description{\packageDescription{autodiff2r}}
\section{Package Content}{\packageIndices{autodiff2r}}
\author{\packageAuthor{autodiff2r}}
\section{Maintainer}{\packageMaintainer{autodiff2r}}
\keyword{package}
