% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autodiff_single_var}
\alias{autodiff_single_var}
\title{autodiff_single_var}
\usage{
autodiff_single_var(input)
}
\arguments{
\item{input}{Value of independent variable (must be greater than 0)}
}
\description{
Example function to show differentiation w.r.t. a single variable
}
